Fix destructuring with defaults - use null context for Assign nodes
The issue was that Param.eachName expects Assign nodes for destructuring
defaults to have null context. We were setting context='=' which caused
the wrong branch to be taken, extracting the value instead of the variable.

This fixes the '1' can't be assigned error in 20-advanced-patterns.test.coffee
and allows the entire file to compile and run.

Test results improved from 365 to 366 passing (96.5%).